backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymongo==4.6.3
python-dotenv==1.0.1
feedparser==6.0.11
beautifulsoup4==4.12.3
google-generativeai==0.5.4
requests==2.31.0
apscheduler==3.10.4
httpx==0.27.0
pydantic==2.6.4
